move Python font bootstrapping artifacts into graphics crate
